Move parser exceptions to a new exceptions package inside parser package
