+++ app/src/main/java/com/example/hwutimetable/parser/exceptions/InvalidRangeException.kt
@@ -1,3 +1,3 @@
-package com.example.hwutimetable.parser
+package com.example.hwutimetable.parser.exceptions
 
 class InvalidRangeException(message: String): Exception(message)
