+++ app/src/main/java/com/example/hwutimetable/parser/exceptions/ParserException.kt
@@ -1,4 +1,4 @@
-package com.example.hwutimetable.parser
+package com.example.hwutimetable.parser.exceptions
 
 import java.lang.Exception
